remove classes which are alternative names of EInt/EShort

+++ fr.centralesupelec.edf.riseclipse.iec61850.nsd/src/fr/centralesupelec/edf/riseclipse/iec61850/nsd/Changes.java
@@ -111,13 +111,13 @@ public interface Changes extends EObject {
      * @return the value of the '<em>Release</em>' attribute.
      * @see #isSetRelease()
      * @see #unsetRelease()
-     * @see #setRelease(short)
+     * @see #setRelease(Integer)
      * @see fr.centralesupelec.edf.riseclipse.iec61850.nsd.NsdPackage#getChanges_Release()
-     * @model default="1" unsettable="true" dataType="fr.centralesupelec.edf.riseclipse.iec61850.nsd.NSRelease"
+     * @model default="1" unsettable="true"
      *        extendedMetaData="kind='attribute' name='release'"
      * @generated
      */
-    short getRelease();
+    Integer getRelease();
 
     /**
      * Sets the value of the '{@link fr.centralesupelec.edf.riseclipse.iec61850.nsd.Changes#getRelease <em>Release</em>}' attribute.
@@ -129,7 +129,7 @@ public interface Changes extends EObject {
      * @see #getRelease()
      * @generated
      */
-    void setRelease( short value );
+    void setRelease( Integer value );
 
     /**
      * Unsets the value of the '{@link fr.centralesupelec.edf.riseclipse.iec61850.nsd.Changes#getRelease <em>Release</em>}' attribute.
@@ -137,7 +137,7 @@ public interface Changes extends EObject {
      * <!-- end-user-doc -->
      * @see #isSetRelease()
      * @see #getRelease()
-     * @see #setRelease(short)
+     * @see #setRelease(Integer)
      * @generated
      */
     void unsetRelease();
@@ -149,7 +149,7 @@ public interface Changes extends EObject {
      * @return whether the value of the '<em>Release</em>' attribute is set.
      * @see #unsetRelease()
      * @see #getRelease()
-     * @see #setRelease(short)
+     * @see #setRelease(Integer)
      * @generated
      */
     boolean isSetRelease();
@@ -243,13 +243,13 @@ public interface Changes extends EObject {
      * @return the value of the '<em>Version</em>' attribute.
      * @see #isSetVersion()
      * @see #unsetVersion()
-     * @see #setVersion(int)
+     * @see #setVersion(Integer)
      * @see fr.centralesupelec.edf.riseclipse.iec61850.nsd.NsdPackage#getChanges_Version()
-     * @model unsettable="true" dataType="fr.centralesupelec.edf.riseclipse.iec61850.nsd.NSVersion" required="true"
+     * @model unsettable="true" required="true"
      *        extendedMetaData="kind='attribute' name='version'"
      * @generated
      */
-    int getVersion();
+    Integer getVersion();
 
     /**
      * Sets the value of the '{@link fr.centralesupelec.edf.riseclipse.iec61850.nsd.Changes#getVersion <em>Version</em>}' attribute.
@@ -261,7 +261,7 @@ public interface Changes extends EObject {
      * @see #getVersion()
      * @generated
      */
-    void setVersion( int value );
+    void setVersion( Integer value );
 
     /**
      * Unsets the value of the '{@link fr.centralesupelec.edf.riseclipse.iec61850.nsd.Changes#getVersion <em>Version</em>}' attribute.
@@ -269,7 +269,7 @@ public interface Changes extends EObject {
      * <!-- end-user-doc -->
      * @see #isSetVersion()
      * @see #getVersion()
-     * @see #setVersion(int)
+     * @see #setVersion(Integer)
      * @generated
      */
     void unsetVersion();
@@ -281,7 +281,7 @@ public interface Changes extends EObject {
      * @return whether the value of the '<em>Version</em>' attribute is set.
      * @see #unsetVersion()
      * @see #getVersion()
-     * @see #setVersion(int)
+     * @see #setVersion(Integer)
      * @generated
      */
     boolean isSetVersion();
