+++ fr.centralesupelec.edf.riseclipse.iec61850.nsd/src/fr/centralesupelec/edf/riseclipse/iec61850/nsd/ServiceNS.java
@@ -357,13 +357,13 @@ public interface ServiceNS extends Copyrighted {
      * @return the value of the '<em>Release</em>' attribute.
      * @see #isSetRelease()
      * @see #unsetRelease()
-     * @see #setRelease(short)
+     * @see #setRelease(Integer)
      * @see fr.centralesupelec.edf.riseclipse.iec61850.nsd.NsdPackage#getServiceNS_Release()
-     * @model default="1" unsettable="true" dataType="fr.centralesupelec.edf.riseclipse.iec61850.nsd.NSRelease"
+     * @model default="1" unsettable="true"
      *        extendedMetaData="kind='attribute' name='release'"
      * @generated
      */
-    short getRelease();
+    Integer getRelease();
 
     /**
      * Sets the value of the '{@link fr.centralesupelec.edf.riseclipse.iec61850.nsd.ServiceNS#getRelease <em>Release</em>}' attribute.
@@ -375,7 +375,7 @@ public interface ServiceNS extends Copyrighted {
      * @see #getRelease()
      * @generated
      */
-    void setRelease( short value );
+    void setRelease( Integer value );
 
     /**
      * Unsets the value of the '{@link fr.centralesupelec.edf.riseclipse.iec61850.nsd.ServiceNS#getRelease <em>Release</em>}' attribute.
@@ -383,7 +383,7 @@ public interface ServiceNS extends Copyrighted {
      * <!-- end-user-doc -->
      * @see #isSetRelease()
      * @see #getRelease()
-     * @see #setRelease(short)
+     * @see #setRelease(Integer)
      * @generated
      */
     void unsetRelease();
@@ -395,7 +395,7 @@ public interface ServiceNS extends Copyrighted {
      * @return whether the value of the '<em>Release</em>' attribute is set.
      * @see #unsetRelease()
      * @see #getRelease()
-     * @see #setRelease(short)
+     * @see #setRelease(Integer)
      * @generated
      */
     boolean isSetRelease();
@@ -516,13 +516,13 @@ public interface ServiceNS extends Copyrighted {
      * @return the value of the '<em>Version</em>' attribute.
      * @see #isSetVersion()
      * @see #unsetVersion()
-     * @see #setVersion(int)
+     * @see #setVersion(Integer)
      * @see fr.centralesupelec.edf.riseclipse.iec61850.nsd.NsdPackage#getServiceNS_Version()
-     * @model unsettable="true" dataType="fr.centralesupelec.edf.riseclipse.iec61850.nsd.NSVersion" required="true"
+     * @model unsettable="true" required="true"
      *        extendedMetaData="kind='attribute' name='version'"
      * @generated
      */
-    int getVersion();
+    Integer getVersion();
 
     /**
      * Sets the value of the '{@link fr.centralesupelec.edf.riseclipse.iec61850.nsd.ServiceNS#getVersion <em>Version</em>}' attribute.
@@ -534,7 +534,7 @@ public interface ServiceNS extends Copyrighted {
      * @see #getVersion()
      * @generated
      */
-    void setVersion( int value );
+    void setVersion( Integer value );
 
     /**
      * Unsets the value of the '{@link fr.centralesupelec.edf.riseclipse.iec61850.nsd.ServiceNS#getVersion <em>Version</em>}' attribute.
@@ -542,7 +542,7 @@ public interface ServiceNS extends Copyrighted {
      * <!-- end-user-doc -->
      * @see #isSetVersion()
      * @see #getVersion()
-     * @see #setVersion(int)
+     * @see #setVersion(Integer)
      * @generated
      */
     void unsetVersion();
@@ -554,7 +554,7 @@ public interface ServiceNS extends Copyrighted {
      * @return whether the value of the '<em>Version</em>' attribute is set.
      * @see #unsetVersion()
      * @see #getVersion()
-     * @see #setVersion(int)
+     * @see #setVersion(Integer)
      * @generated
      */
     boolean isSetVersion();
